+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 - Essential steps
 - Commands & flags structure (cmdline)
 - CLI steps for toolchain development
+
+## 1.0.1
+- Added ``onFailure()`` & ``onSuccess()`` to the ``Check``-step.
+- Fixed the output of ``Check`` when ``canStartProcesses`` is set to ``false``.
+- Added ``Check`` unit test.